Split extensions into different files
