+++ src/ComputeSharp.SourceGeneration/Extensions/ISymbolExtensions.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Linq;
-using ComputeSharp.SourceGeneration.Helpers;
 using Microsoft.CodeAnalysis;
 
 namespace ComputeSharp.SourceGeneration.Extensions;
@@ -27,142 +24,6 @@ internal static class ISymbolExtensions
     public static string GetFullyQualifiedName(this ISymbol symbol)
     {
         return symbol.ToDisplayString(FullyQualifiedWithoutGlobalFormat);
-    }
-
-    /// <summary>
-    /// Checks whether or not a given type symbol has a specified fully qualified metadata name.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="ITypeSymbol"/> instance to check.</param>
-    /// <param name="name">The full name to check.</param>
-    /// <returns>Whether <paramref name="symbol"/> has a full name equals to <paramref name="name"/>.</returns>
-    public static bool HasFullyQualifiedMetadataName(this ITypeSymbol symbol, string name)
-    {
-        using ImmutableArrayBuilder<char> builder = ImmutableArrayBuilder<char>.Rent();
-
-        AppendFullyQualifiedMetadataName(symbol, in builder);
-
-        return builder.WrittenSpan.SequenceEqual(name.AsSpan());
-    }
-
-    /// <summary>
-    /// Gets the fully qualified metadata name for a given <see cref="ITypeSymbol"/> instance.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="ITypeSymbol"/> instance.</param>
-    /// <returns>The fully qualified metadata name for <paramref name="symbol"/>.</returns>
-    public static string GetFullyQualifiedMetadataName(this ITypeSymbol symbol)
-    {
-        using ImmutableArrayBuilder<char> builder = ImmutableArrayBuilder<char>.Rent();
-
-        AppendFullyQualifiedMetadataName(symbol, in builder);
-
-        return builder.ToString();
-    }
-
-    /// <summary>
-    /// Gets the fully qualified metadata name for a given <see cref="IMethodSymbol"/> instance.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="IMethodSymbol"/> instance.</param>
-    /// <param name="includeParameters">Whether or not to also include parameter types.</param>
-    /// <returns>The fully qualified metadata name for <paramref name="symbol"/>.</returns>
-    public static string GetFullyQualifiedMetadataName(this IMethodSymbol symbol, bool includeParameters = false)
-    {
-        using ImmutableArrayBuilder<char> builder = ImmutableArrayBuilder<char>.Rent();
-
-        AppendFullyQualifiedMetadataName(symbol.ContainingType!, in builder);
-
-        // Always add the method name after the containing type
-        builder.Add('.');
-        builder.AddRange(symbol.Name.AsSpan());
-
-        if (includeParameters)
-        {
-            builder.Add('(');
-
-            bool isFirstParameter = true;
-
-            foreach (IParameterSymbol parameter in symbol.Parameters)
-            {
-                // Append the leading ', ' if needed
-                if (isFirstParameter)
-                {
-                    isFirstParameter = false;
-                }
-                else
-                {
-                    builder.AddRange(", ".AsSpan());
-                }
-
-                AppendFullyQualifiedMetadataName(parameter.Type, in builder);
-            }
-
-            builder.Add(')');
-        }
-
-        return builder.ToString();
-    }
-
-    /// <summary>
-    /// Gets the fully qualified metadata name for a given <see cref="IPropertySymbol"/> instance.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="IPropertySymbol"/> instance.</param>
-    /// <returns>The fully qualified metadata name for <paramref name="symbol"/>.</returns>
-    public static string GetFullyQualifiedMetadataName(this IPropertySymbol symbol)
-    {
-        using ImmutableArrayBuilder<char> builder = ImmutableArrayBuilder<char>.Rent();
-
-        AppendFullyQualifiedMetadataName(symbol.ContainingType!, in builder);
-
-        if (symbol.IsIndexer)
-        {
-            // Add the ".this[...]" suffix (parameters will be added below)
-            builder.AddRange(".this[".AsSpan());
-
-            bool isFirstParameter = true;
-
-            // If the property is an indexer, also includes the parameters
-            foreach (IParameterSymbol parameter in symbol.Parameters)
-            {
-                // Append the leading ', ' if needed
-                if (isFirstParameter)
-                {
-                    isFirstParameter = false;
-                }
-                else
-                {
-                    builder.AddRange(", ".AsSpan());
-                }
-
-                AppendFullyQualifiedMetadataName(parameter.Type, in builder);
-            }
-
-            // Add the closing bracket for ".this[...]"
-            builder.Add(']');
-        }
-        else
-        {
-            // For normal properties, just append the name after the containing type
-            builder.Add('.');
-            builder.AddRange(symbol.Name.AsSpan());
-        }
-
-        return builder.ToString();
-    }
-
-    /// <summary>
-    /// Gets the fully qualified metadata name for a given <see cref="IFieldSymbol"/> instance.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="IFieldSymbol"/> instance.</param>
-    /// <returns>The fully qualified metadata name for <paramref name="symbol"/>.</returns>
-    public static string GetFullyQualifiedMetadataName(this IFieldSymbol symbol)
-    {
-        using ImmutableArrayBuilder<char> builder = ImmutableArrayBuilder<char>.Rent();
-
-        AppendFullyQualifiedMetadataName(symbol.ContainingType!, in builder);
-
-        builder.Add('.');
-        builder.AddRange(symbol.Name.AsSpan());
-
-        return builder.ToString();
     }
 
     /// <summary>
@@ -241,54 +102,5 @@ internal static class ISymbolExtensions
         return
             accessibility == Accessibility.Public ||
             (accessibility == Accessibility.Internal && symbol.ContainingAssembly.GivesAccessTo(assembly));
-    }
-
-    /// <summary>
-    /// Appends the fully qualified metadata name for a given symbol to a target builder.
-    /// </summary>
-    /// <param name="symbol">The input <see cref="ITypeSymbol"/> instance.</param>
-    /// <param name="builder">The target <see cref="ImmutableArrayBuilder{T}"/> instance.</param>
-    private static void AppendFullyQualifiedMetadataName(this ITypeSymbol symbol, in ImmutableArrayBuilder<char> builder)
-    {
-        static void BuildFrom(ISymbol? symbol, in ImmutableArrayBuilder<char> builder)
-        {
-            switch (symbol)
-            {
-                // Namespaces that are nested also append a leading '.'
-                case INamespaceSymbol { ContainingNamespace.IsGlobalNamespace: false }:
-                    BuildFrom(symbol.ContainingNamespace, in builder);
-                    builder.Add('.');
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-
-                // Other namespaces (ie. the one right before global) skip the leading '.'
-                case INamespaceSymbol { IsGlobalNamespace: false }:
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-
-                // Types with no namespace just have their metadata name directly written
-                case ITypeSymbol { ContainingSymbol: INamespaceSymbol { IsGlobalNamespace: true } }:
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-
-                // Types with a containing non-global namespace also append a leading '.'
-                case ITypeSymbol { ContainingSymbol: INamespaceSymbol namespaceSymbol }:
-                    BuildFrom(namespaceSymbol, in builder);
-                    builder.Add('.');
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-
-                // Nested types append a leading '+'
-                case ITypeSymbol { ContainingSymbol: ITypeSymbol typeSymbol }:
-                    BuildFrom(typeSymbol, in builder);
-                    builder.Add('+');
-                    builder.AddRange(symbol.MetadataName.AsSpan());
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        BuildFrom(symbol, in builder);
     }
 }
